feat: Add source button to `/word`.
